CCR-640
8.2.1 should change field name to: Is there a signed delegation of
authority between the UNDP Resident Representative and the UNDP Country
Director and/or UNDP Deputy Country Director?

9.3,11.7, 12.7.1.3 shouldn't appear on Saiku

14.2.A should be on Saiku but it doesn't appear.

Can you please add 12.3.2 as it is now a Y/N question.

For 6.3.A while I see it all there in Saiku, when I select it as a
measure, it says 0, even for those that said yes. The Dimension for that
question seems to work ok.